requirements.txt
aiohttp==3.6.3
async-timeout==3.0.1
attrs==20.3.0
boto==2.49.0
chardet==3.0.4
discord.py==1.5.1
idna==2.10
MouseInfo==0.1.3
multidict==4.7.6
numpy==1.19.3
opencv-python==4.4.0.46
Pillow==8.0.1
pip==20.2.4
PyAutoGUI==0.9.52
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.1
PyRect==0.1.4
PyScreeze==0.1.26
pytesseract==0.3.6
python-dotenv==0.15.0
PyTweening==1.0.3
setuptools==47.1.0
typing-extensions==3.7.4.3
yarl==1.5.1
